+++ spring-telqos-api/src/main/java/com/dwarfeng/springtelqos/api/integration/system/JmxRemoteCommand.java
@@ -79,10 +79,10 @@ public class JmxRemoteCommand extends CliCommand {
 
     private String cliSyntaxProvider(CommandDescriptor.Context context) throws Exception {
         final String[] patterns = new String[]{
-                context.getIdentity() + " " + CliCommandUtil.concatOptionPrefix(COMMAND_OPTION_START) + " [" +
+                context.getRuntimeIdentity() + " " + CliCommandUtil.concatOptionPrefix(COMMAND_OPTION_START) + " [" +
                         CliCommandUtil.concatOptionPrefix(COMMAND_SUB_OPTION_PORT) + " port]",
-                context.getIdentity() + " " + CliCommandUtil.concatOptionPrefix(COMMAND_OPTION_STOP),
-                context.getIdentity() + " " + CliCommandUtil.concatOptionPrefix(COMMAND_OPTION_STATUS)
+                context.getRuntimeIdentity() + " " + CliCommandUtil.concatOptionPrefix(COMMAND_OPTION_STOP),
+                context.getRuntimeIdentity() + " " + CliCommandUtil.concatOptionPrefix(COMMAND_OPTION_STATUS)
         };
         return CliCommandUtil.cliSyntax(patterns);
     }
@@ -102,7 +102,7 @@ public class JmxRemoteCommand extends CliCommand {
         Pair<String, Integer> pair = CliCommandUtil.analyseCommand(cmd, COMMAND_OPTION_ARRAY);
         if (pair.getRight() != 1) {
             context.sendMessage(CliCommandUtil.optionMismatchMessage(COMMAND_OPTION_ARRAY));
-            context.sendMessage(context.getCommandManual(context.getIdentity()));
+            context.sendMessage(context.getCommandManual(context.getRuntimeIdentity()));
             return;
         }
         switch (pair.getLeft()) {